+++ scripts/__init__.py
@@ -5,4 +5,4 @@
 
 """sny copyright check pre-commit hook"""
 
-__version__ = "1.0.2"
+__version__ = "1.0.3"